fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@24388 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/div/GUIParameterTableWindow.cpp
@@ -272,7 +272,7 @@ GUIParameterTableWindow::closeBuilding(const Parameterised* p) {
 int 
 GUIParameterTableWindow::numParams(const GUIGlObject* obj) {
     const Parameterised* p = dynamic_cast<const Parameterised*>(obj);
-    return p != 0 ? p->getMap().size() : 0;
+    return p != 0 ? (int)p->getMap().size() : 0;
 }
